==> 1_preprocessing.ipynb <==
# Bots in Science 🧪

In this notebook, we process data from mentions to:
+ (1) classify altmetric papers by ESI field 📚
+ (2) analyze the botscores of the tweeters to identify bots 🤖 in combination with their activity 🕵️‍♂️

## Libraries

import pandas as pd
import numpy as np
from functions import outliers

## 1. Altmetric papers by ESI field

The Web of Science papers (2017-2021) indexed in altmetric are assigned one or more Web of Science categories. Based on the schema by Arroyo-Machado & Torres-Salinas (2021), their ESI field is obtained.

データ_alt_sub = pd.read_csv('data/altmetric_subjects_cat.tsv', sep='\t')
データ_alt_sub.DOI = データ_alt_sub.DOI.str.lower()
データ_alt_sub.shape

データ_esi = pd.read_csv('data/mapping.csv', sep=';')
データ_esi

データ_alt_sub = データ_alt_sub.merge(データ_esi[['WC', 'ESI']], how='inner', left_on='subject_category', right_on='WC')
データ_alt_sub.drop('WC', axis=1, inplace=True)

<div class="alert-warning">
    <strong>Warning:</strong> This line of code is commented to avoid generating new versions of the file when reviewing the code.
</div>

#データ_alt_sub.to_csv('data/altmetric_subjects_cat_esi.tsv', index=False, sep='\t', encoding='UTF-8')

## 2. Bot detection

Based on the botscore and the frequency of mentions of the tweeters, bots are identified.

データ_botscore = pd.read_csv('data/full_botometer_results.tsv', sep='\t', dtype={'user_id':str})
データ_botscore.shape

データ_tw_men = pd.read_csv('data/final_mentions_full.tsv', sep='\t',
                         dtype={'Outlet or Author':str, 'External Mention ID':str},
                         encoding='UTF-8')
データ_tw_men.shape

Out of the 4,983,251 tweeters, it has been possible to obtain the botscore of 4,872,369 tweeters (98%).

データ_tw_men[['Outlet or Author']].drop_duplicates().shape

データ_tw_men[データ_tw_men['Outlet or Author'].isin(データ_botscore.user_id.tolist())][['Outlet or Author']].drop_duplicates().shape

62 publications don't include Web of Science category. Possibly they were not correctly indexed when the data was retrieved.

len(データ_tw_men.loc[~データ_tw_men.DOI.isin(データ_alt_sub.DOI.tolist()), 'DOI'].drop_duplicates())

### 2.1. Outliers detection

In order to identify the bots, user mentions are taken into account. Therefore, the outliers of this distribution (IQR*1.5 & IQR*3) are calculated.

### 2.1.1. Mentions

データ_tw_men_freq = データ_tw_men[['Outlet or Author']].value_counts().reset_index(name='mentions')
データ_tw_men_freq

Outliers realize more than 8.5 mentions and extreme outliers more than 13.

outliers(データ_tw_men_freq.mentions)

outliers(データ_tw_men_freq.mentions, extreme=1)

### 2.1.2. Tweets

データ_tw_men_freq = データ_tw_men.groupby('Outlet or Author').sum('Original').reset_index()
データ_tw_men_freq.rename({'Original':'tweets'}, axis=1, inplace=True)
データ_tw_men_freq

Outliers realize more than 2.5 tweets and extreme outliers more than 4.

outliers(データ_tw_men_freq.tweets)

outliers(データ_tw_men_freq.tweets, extreme=1)

### 2.1.3. Retweets

データ_tw_men_freq = データ_tw_men
データ_tw_men_freq.Original = -1*(データ_tw_men_freq.Original-1)
データ_tw_men_freq = データ_tw_men_freq.groupby('Outlet or Author').sum('Original').reset_index()
データ_tw_men_freq.rename({'Original':'retweets'}, axis=1, inplace=True)
データ_tw_men_freq

Outliers realize more than 6 retweets and extreme outliers more than 9.

outliers(データ_tw_men_freq.retweets)

outliers(データ_tw_men_freq.retweets, extreme=1)

### 2.2. Bots detection

After some tests that are available on the robustenss check, it has been decided to set as thresholds a botscores of 0.6 and a minimum of 8 mentions.

データ_tw_men_freq = データ_tw_men[['Outlet or Author']].value_counts().reset_index(name='mentions')
データ_tw_men_freq = データ_tw_men_freq[データ_tw_men_freq.mentions>8]
データ_tw_men_freq

データ_bots = データ_tw_men_freq.merge(データ_botscore, how='inner', left_on='Outlet or Author', right_on='user_id')
データ_bots = データ_bots[データ_bots.botscore>0.6].copy()
データ_bots = データ_bots[['Outlet or Author']].copy()
データ_bots

<div class="alert-warning">
    <strong>Warning:</strong> This line of code is commented to avoid generating new versions of the file when reviewing the code.
</div>

#データ_bots.to_csv('results/bots_list.tsv', index=False, sep='\t')

# References

Arroyo-Machado, W., & Torres-Salinas, D. (2021). *Web of Science categories (WC, SC, main categories) and ESI disciplines mapping*. https://doi.org/10.6084/m9.figshare.14695176.v2
