==> 2_bots_mentions.ipynb <==
# Bots in Science 🧪

In this notebook, mentions made by bots are identified:
+ (1) globally 🌍
+ (2) by Web of Science category 📖
+ (3) by ESI field 📚

## Libraries

import pandas as pd

## 1. Overall bots

### 1.1. Preprocessing

Mentions made by tweeters for whom a botscore could be estimated are filtered (98.5%).

データ_tw_men = pd.read_csv('data/final_mentions_full.tsv', sep='\t',
                         dtype={'Outlet or Author':str, 'External Mention ID':str},
                         encoding='UTF-8')
データ_tw_men.shape

データ_botscore = pd.read_csv('data/full_botometer_results.tsv', sep='\t', dtype={'user_id':str})
データ_botscore.shape

データ_tw_men = データ_tw_men[データ_tw_men['Outlet or Author'].isin(データ_botscore.user_id.tolist())].copy()
データ_tw_men.shape

### 1.2. Bots

The mentions are labeled according to whether the tweeter making them is a bot or not.

データ_bots = pd.read_csv('results/bots_list.tsv', sep='\t', dtype={'Outlet or Author':str})
データ_bots.shape

データ_tw_men['bot'] = 0
データ_tw_men.loc[データ_tw_men['Outlet or Author'].isin(データ_bots['Outlet or Author'].tolist()), 'bot'] = 1

データ_tw_men[['bot']].value_counts()/データ_tw_men.shape[0]

<div class="alert-warning">
    <strong>Warning:</strong> This line of code is commented to avoid generating new versions of the file when reviewing the code.
</div>

#データ_tw_men.to_csv('data/final_mentions_full_bots.tsv', index=False, sep='\t', encoding='UTF-8')

## 2. Bots by category

As above but classifying mentions by Web of Science category.

データ_alt_sub = pd.read_csv('data/altmetric_subjects_cat_esi.tsv', sep='\t', encoding='UTF-8')
データ_alt_sub

データ_tw_men.DOI = データ_tw_men.DOI.str.lower()
データ_alt_sub.DOI = データ_alt_sub.DOI.str.lower()

データ_alt_sub_cat = データ_alt_sub[['DOI', 'subject_category']].drop_duplicates()
データ_alt_sub_cat.shape

Due to the lack of categories in 64 publications, 524 mentions are lost.

データ_tw_men[データ_tw_men.DOI.isin(データ_alt_sub.DOI.tolist())].shape

データ_tw_men_cat = データ_tw_men.merge(データ_alt_sub_cat, how='inner', on='DOI')
データ_tw_men_cat.shape

<div class="alert-warning">
    <strong>Warning:</strong> This line of code is commented to avoid generating new versions of the file when reviewing the code.
</div>

#データ_tw_men_cat.to_csv('data/final_mentions_full_bots_category.tsv', index=False, sep='\t', encoding='UTF-8')

## 3. Bots by ESI field

データ_alt_sub_esi = データ_alt_sub[['DOI', 'ESI']].drop_duplicates()
データ_alt_sub_esi.shape

データ_tw_men_esi = データ_tw_men.merge(データ_alt_sub_esi, how='inner', on='DOI')
データ_tw_men_esi.shape

<div class="alert-warning">
    <strong>Warning:</strong> This line of code is commented to avoid generating new versions of the file when reviewing the code.
</div>

#データ_tw_men_esi.to_csv('data/final_mentions_full_bots_esi.tsv', index=False, sep='\t', encoding='UTF-8')
