==> 3_bots_mentions_stats.ipynb <==
# Bots in Science 🧪

In this notebook, the presence statistics of bots are calculated in:
+ (1) general 🌍
+ (2) by Web of Science category (Mathematics is individually explored) 📖
+ (3) by ESI field 📚

## Libraries

import pandas as pd
from functions import bot_mentions

## 1. Overall

データ_tw_men = pd.read_csv('data/final_mentions_full_bots.tsv', sep='\t', encoding='UTF-8',
                         dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men.shape

データ_tw_men_freq = bot_mentions(データ_tw_men)
データ_tw_men_freq

#データ_tw_men_freq.to_csv('results/agg_bot_mentions.tsv', sep='\t', index=False)

## 2. Web of Science categories

データ_tw_men_cat = pd.read_csv('data/final_mentions_full_bots_category.tsv', sep='\t', encoding='UTF-8',
                             dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men_cat.shape

データ_tw_men_cat_freq = bot_mentions(データ_tw_men_cat, agg_by='subject_category')
データ_tw_men_cat_freq

<div class="alert-warning">
    <strong>Warning:</strong> This line of code is commented to avoid generating new versions of the file when reviewing the code.
</div>

#データ_tw_men_cat_freq.to_csv('results/agg_bot_mentions_cat.tsv', sep='\t', index=True)

### 2.1. Mathematics

Due to the high number of bot mentions and their concentration in a few accounts, Mathematics has been explored.

データ_tw_men_cat_freq.loc['Mathematics']

データ_tweeters = pd.read_csv('data/tweeters_metadata.tsv', sep='\t', encoding='UTF-8', dtype={'id_str':str})
データ_tweeters = データ_tweeters.groupby('id_str').first().reset_index()
データ_tweeters.shape

データ_mat = データ_tw_men_cat[データ_tw_men_cat['subject_category']=='Mathematics']
データ_mat.shape

データ_mat_bot = データ_mat[データ_mat.bot==1].copy()
データ_mat_bot = bot_mentions(データ_mat_bot, agg_by='Outlet or Author')

sum(データ_mat_bot.mentions)

データ_mat_bot = データ_mat[データ_mat.bot==1].copy()
データ_mat_bot = bot_mentions(データ_mat_bot, agg_by='Outlet or Author')
データ_mat_bot = データ_mat_bot.reset_index()
データ_mat_bot.shape

データ_mat_bot = データ_mat_bot.merge(データ_tweeters, how='inner', left_on='Outlet or Author', right_on='id_str')
データ_mat_bot.shape

<div class="alert-warning">
    <strong>Warning:</strong> This line of code is commented to avoid generating new versions of the file when reviewing the code.
</div>

#データ_mat_bot.to_csv('results/mathematics.tsv', sep='\t', index=False, encoding='UTF-8')

## 3. ESI field

データ_tw_men_esi = pd.read_csv('data/final_mentions_full_bots_esi.tsv', sep='\t', encoding='UTF-8',
                             dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men_esi.shape

データ_tw_men_esi_freq = bot_mentions(データ_tw_men_esi, agg_by='ESI')
データ_tw_men_esi_freq

<div class="alert-warning">
    <strong>Warning:</strong> This line of code is commented to avoid generating new versions of the file when reviewing the code.
</div>

#データ_tw_men_esi_freq.to_csv('results/agg_bot_mentions_esi.tsv', sep='\t', index=True)

データ_tw_men_cat[(データ_tw_men_cat['subject_category']=='Literature, Slavic') & (データ_tw_men_cat.bot==1)].sort_values('Outlet or Author')
