==> 4_bot_profiles.ipynb <==
# Bots in Science 🧪

## Libraries

import pandas as pd
from functions import bot_mentions

## 1. Overall

データ_bots = pd.read_csv('results/bots_list.tsv', sep='\t', dtype={'Outlet or Author':str})
データ_bots.shape

データ_tweeters = pd.read_csv('data/tweeters_metadata.tsv', sep='\t', encoding='UTF-8', dtype={'id_str':str})
データ_tweeters = データ_tweeters.groupby('id_str').first().reset_index()
データ_tweeters = データ_tweeters[データ_tweeters.id_str.isin(データ_bots['Outlet or Author'].tolist())].copy()
データ_tweeters.shape

データ_tw_men = pd.read_csv('data/final_mentions_full_bots.tsv', sep='\t', encoding='UTF-8',
                         dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men = データ_tw_men[データ_tw_men['Outlet or Author'].isin(データ_bots['Outlet or Author'].tolist())].copy()
データ_tw_men.shape

データ_tw_men_bot_freq = bot_mentions(データ_tw_men, agg_by='Outlet or Author')
データ_tw_men_bot_freq = データ_tw_men_bot_freq.reset_index()
データ_tw_men_bot_freq

データ_tw_men_bot_freq = データ_bots.merge(データ_tw_men_bot_freq, how='inner', on='Outlet or Author')
データ_tw_men_bot_freq

データ_tw_men_bot_freq = データ_tw_men_bot_freq[['Outlet or Author', 'papers_mentioned', 'mentions', 'tweets', 'retweet']].copy()
データ_tw_men_bot_freq = データ_tw_men_bot_freq.merge(データ_tweeters, how='inner', left_on='Outlet or Author', right_on='id_str')

データ_tw_men_bot_freq[['papers_mentioned', 'mentions', 'tweets', 'retweet']].groupby(lambda x: True).agg({'mean', 'median', 'std'})

#データ_tw_men_bot_freq.sort_values('mentions', ascending=False).head(30).to_csv('results/top_bots.tsv', sep='\t', encoding='UTF-8', index=False)

データ_tw_men_esi = pd.read_csv('data/final_mentions_full_bots_esi.tsv', sep='\t', encoding='UTF-8',
                             dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men_esi = データ_tw_men_esi[データ_tw_men_esi['Outlet or Author'].isin(データ_bots['Outlet or Author'].tolist())].copy()
データ_tw_men_esi.shape

データ_tw_men_esi_freq = データ_tw_men_esi[['Outlet or Author', 'ESI']].drop_duplicates().groupby('Outlet or Author').count().reset_index()

100*データ_tw_men_esi_freq[データ_tw_men_esi_freq['ESI']>1].shape[0]/データ_tw_men_esi_freq.shape[0]

データ_tw_men_esi_freq['ESI'].mean()
