==> 5_bots_mentions_papers_stats.ipynb <==
# Bots in Science 🧪

In this notebook, the impact of bots on tweets is calculated at the paper level, differentiating in:
+ overall 🌍
+ by ESI field 📚
+ by Web of Science category 📖

## Libraries

import pandas as pd
import numpy as np
from functions import bot_mentions, gini
import scipy.stats

## 1. Overall

In this case, by focusing attention on the tweets and wanting to analyze the impact, the tweets are selected. After that, bot tweets are calculated at the paper level, and metrics are calculated to shed light on the impact of bots on this altmetric.

データ_tw_men = pd.read_csv('data/final_mentions_full_bots.tsv', sep='\t', encoding='UTF-8',
                         dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men = データ_tw_men[データ_tw_men.Original==1].copy()
データ_tw_men.shape

データ_tw_men_paper = bot_mentions(データ_tw_men, agg_by='DOI')

### 1.1. Impact metrics

データ_tw_men_paper['bot_tweets_p'].mean()

gini(np.array(データ_tw_men_paper['bot_tweets_p']))

100*データ_tw_men_paper[データ_tw_men_paper['bot_tweets_p']==100].shape[0]/データ_tw_men_paper.shape[0]

100*データ_tw_men_paper[データ_tw_men_paper['bot_tweets_p']>=50].shape[0]/データ_tw_men_paper.shape[0]

100*データ_tw_men_paper[データ_tw_men_paper['bot_tweets_p']>0].shape[0]/データ_tw_men_paper.shape[0]

データ_tw_men_paper['tweets_no_bot'] = データ_tw_men_paper['tweets']-データ_tw_men_paper['tweets_bot']
データ_tw_men_paper['tweets_per'] = データ_tw_men_paper['tweets'].rank(pct=True)
データ_tw_men_paper['tweets_no_bot_per'] = データ_tw_men_paper['tweets_no_bot'].rank(pct=True)
データ_tw_men_paper['per_diff'] = abs(データ_tw_men_paper['tweets_per']-データ_tw_men_paper['tweets_no_bot_per'])

100*データ_tw_men_paper['per_diff'].mean()

scipy.stats.spearmanr(データ_tw_men_paper['tweets'], データ_tw_men_paper['tweets_no_bot'])[0]

## 2. ESI field

The same process is replicated but at the ESI field level.

データ_tw_men_esi = pd.read_csv('data/final_mentions_full_bots_esi.tsv', sep='\t', encoding='UTF-8',
                             dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men_esi = データ_tw_men_esi.loc[データ_tw_men_esi.Original==1, ['ESI', 'DOI']].drop_duplicates().copy()
データ_tw_men_esi.shape

データ_tw_men_esi = データ_tw_men_esi.merge(データ_tw_men_paper, how='inner', on='DOI')

### 2.1. Impact metrics

データ_tw_men_esi['tweets_per'] = None
データ_tw_men_esi['tweets_no_bot_per'] = None
データ_tw_men_esi['per_diff'] = None

for esi in list(set(データ_tw_men_esi['ESI'])):
    データ_tw_men_esi.loc[データ_tw_men_esi.ESI==esi, 'tweets_per'] = 100*データ_tw_men_esi.loc[データ_tw_men_esi.ESI==esi, 'tweets'].rank(pct=True)
    データ_tw_men_esi.loc[データ_tw_men_esi.ESI==esi,'tweets_no_bot_per'] = 100*データ_tw_men_esi.loc[データ_tw_men_esi.ESI==esi, 'tweets_no_bot'].rank(pct=True)
データ_tw_men_esi['per_diff'] = abs(データ_tw_men_esi['tweets_per']-データ_tw_men_esi['tweets_no_bot_per'])

データ_tw_men_esi_paper = データ_tw_men_esi.groupby('ESI').agg({'bot_tweets_p':'mean', 'per_diff':'mean'}).reset_index()

データ_tw_men_esi_paper_aux = データ_tw_men_esi.groupby('ESI').agg({
    'bot_tweets_p': [
        lambda x: 100*(x > 0).sum()/len(x),
        lambda x: 100*(x >= 50).sum()/len(x),
        lambda x: 100*(x == 100).sum()/len(x)
    ]
}).reset_index()
データ_tw_men_esi_paper_aux.columns = ['ESI', 'any', 'half', 'all']

データ_tw_men_esi_paper = データ_tw_men_esi_paper.merge(データ_tw_men_esi_paper_aux, how='inner', on='ESI')

データ_tw_men_esi_gini = データ_tw_men_esi.groupby('ESI').apply(
    lambda x: gini(np.array(x['bot_tweets_p']))
).reset_index().rename({0:'gini'}, axis=1)

データ_tw_men_esi_paper = データ_tw_men_esi_paper.merge(データ_tw_men_esi_gini, how='inner', on='ESI')

データ_tw_men_esi_corr = データ_tw_men_esi.groupby('ESI').apply(
    lambda x: scipy.stats.spearmanr(x['tweets'], x['tweets_no_bot'])[0]
).reset_index().rename({0:'corr'}, axis=1)

データ_tw_men_esi_paper = データ_tw_men_esi_paper.merge(データ_tw_men_esi_corr, how='inner', on='ESI')

データ_tw_men_esi_paper

## 3. Categories

データ_tw_men_cat = pd.read_csv('data/final_mentions_full_bots_category.tsv', sep='\t', encoding='UTF-8',
                             dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men_cat = データ_tw_men_cat[['subject_category', 'DOI']].drop_duplicates().copy()
データ_tw_men_cat.shape

データ_tw_men_cat = データ_tw_men_cat.merge(データ_tw_men_paper, how='inner', on='DOI')

### 3.1. Impact metrics

データ_tw_men_cat['tweets_per'] = None
データ_tw_men_cat['tweets_no_bot_per'] = None
データ_tw_men_cat['per_diff'] = None

for category in list(set(データ_tw_men_cat['subject_category'])):
    データ_tw_men_cat.loc[データ_tw_men_cat.subject_category==category, 'tweets_per'] = 100*データ_tw_men_cat.loc[データ_tw_men_cat.subject_category==category, 'tweets'].rank(pct=True)
    データ_tw_men_cat.loc[データ_tw_men_cat.subject_category==category,'tweets_no_bot_per'] = 100*データ_tw_men_cat.loc[データ_tw_men_cat.subject_category==category, 'tweets_no_bot'].rank(pct=True)
データ_tw_men_cat['per_diff'] = abs(データ_tw_men_cat['tweets_per']-データ_tw_men_cat['tweets_no_bot_per'])

データ_tw_men_cat_paper = データ_tw_men_cat.groupby('subject_category').agg({'bot_tweets_p':'mean', 'per_diff':'mean'}).reset_index()

データ_tw_men_cat_paper_aux = データ_tw_men_cat.groupby('subject_category').agg({
    'bot_tweets_p': [
        lambda x: 100*(x > 0).sum()/len(x),
        lambda x: 100*(x >= 50).sum()/len(x),
        lambda x: 100*(x == 100).sum()/len(x)
    ]
}).reset_index()
データ_tw_men_cat_paper_aux.columns = ['subject_category', 'any', 'half', 'all']

データ_tw_men_cat_paper = データ_tw_men_cat_paper.merge(データ_tw_men_cat_paper_aux, how='inner', on='subject_category')

データ_tw_men_cat_gini = データ_tw_men_cat.groupby('subject_category').apply(
    lambda x: gini(np.array(x['bot_tweets_p']))
).reset_index().rename({0:'gini'}, axis=1)

データ_tw_men_cat_paper = データ_tw_men_cat_paper.merge(データ_tw_men_cat_gini, how='inner', on='subject_category')

データ_tw_men_cat_corr = データ_tw_men_cat.groupby('subject_category').apply(
    lambda x: scipy.stats.spearmanr(x['tweets'], x['tweets_no_bot'])[0]
).reset_index().rename({0:'corr'}, axis=1)

データ_tw_men_cat_paper = データ_tw_men_cat_paper.merge(データ_tw_men_cat_corr, how='inner', on='subject_category')

データ_tw_men_cat_paper.sort_values(0).head(25)

### 3.2. Categories by main areas

データ_esi = pd.read_csv('data/mapping.csv', sep=';')
データ_esi.rename({'WC':'subject_category'}, axis=1, inplace=True)
データ_esi

データ_tw_men_cat_paper_area = データ_tw_men_cat_paper.merge(データ_esi[['subject_category', 'Category']], on='subject_category')

データ_tw_men_cat_paper_area.Category.value_counts()

データ_tw_men_cat_paper_area[データ_tw_men_cat_paper_area.Category=='Technology'].sort_values(0).head(5)
