==> 6_open_access.ipynb <==
# Bots in Science 🧪

## Libraries

import pandas as pd

## 1. Open Access

データ_oa = pd.read_csv('data/oa_papers.tsv', sep='\t', encoding='UTF-8')
データ_oa.doi = データ_oa.doi.str.lower()
データ_oa.open_access_type.fillna('Closed', inplace=True)
データ_oa.shape

データ_oa.open_access_type.value_counts()

# https://webofscience.help.clarivate.com/en-us/Content/open-access.html
# https://peerj.com/articles/4375/?ref=blog.oa.works

データ_oa.loc[データ_oa.open_access_type=='Green published', 'open_access_type'] = 'Green'
データ_oa.loc[データ_oa.open_access_type=='Green submitted', 'open_access_type'] = 'Green'
データ_oa.loc[データ_oa.open_access_type=='Green accepted', 'open_access_type'] = 'Green'
データ_oa = データ_oa.drop_duplicates()
データ_oa.open_access_type.value_counts()

データ_oa_oa = データ_oa[データ_oa.open_access_type.isin(['Gold', 'Green'])].copy()
データ_oa_other = データ_oa[~データ_oa.open_access_type.isin(['Gold', 'Green'])].copy()

データ_oa_oa.shape

データ_oa_oa = データ_oa_oa.sort_values('open_access_type')
データ_oa_oa = データ_oa_oa.groupby(['doi', 'is_open_access']).first().reset_index()

データ_oa_oa.shape

データ_oa = pd.concat([データ_oa_oa, データ_oa_other])

## 1. Overall

データ_tw_men = pd.read_csv('data/final_mentions_full_bots.tsv', sep='\t', encoding='UTF-8',
                         dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men.DOI = データ_tw_men.DOI.str.lower()
データ_tw_men.shape

データ_tw_men_oa = データ_tw_men.merge(データ_oa, how='inner', left_on='DOI', right_on='doi')
データ_tw_men_oa.shape

データ_tw_men_oa = データ_tw_men_oa.loc[データ_tw_men_oa.Original==1, ['DOI', 'bot', 'open_access_type']].drop_duplicates()
データ_tw_men_oa.shape

データ_tw_men_count = データ_tw_men_oa[['DOI', 'bot']].drop_duplicates().groupby('bot').count()
データ_tw_men_count = データ_tw_men_count.pivot_table(values = 'DOI', index='bot').reset_index()
データ_tw_men_oa_count = データ_tw_men_oa.groupby(['bot', 'open_access_type']).count().reset_index()
データ_tw_men_oa_count = データ_tw_men_oa_count.pivot_table(values = 'DOI', index='bot', columns = 'open_access_type').reset_index()
データ_tw_men_oa_count = データ_tw_men_oa_count.merge(データ_tw_men_count, how='inner', on='bot')

データ_tw_men_oa_count

## 2. ESI field

データ_tw_men_esi = pd.read_csv('data/final_mentions_full_bots_esi.tsv', sep='\t', encoding='UTF-8',
                             dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men_esi.DOI = データ_tw_men_esi.DOI.str.lower()
データ_tw_men_esi.shape

データ_tw_men_esi_oa = データ_tw_men_esi.merge(データ_oa, how='inner', left_on='DOI', right_on='doi')
データ_tw_men_esi_oa.shape

データ_tw_men_esi_oa = データ_tw_men_esi_oa.loc[データ_tw_men_esi_oa.Original==1, ['ESI', 'DOI', 'bot', 'open_access_type']].drop_duplicates()
データ_tw_men_esi_oa.shape

データ_tw_men_esi_count = データ_tw_men_esi_oa[['ESI', 'DOI', 'bot']].drop_duplicates().groupby(['ESI', 'bot']).count()
データ_tw_men_esi_count = データ_tw_men_esi_count.pivot_table(values = 'DOI', index=['bot', 'ESI']).reset_index()
データ_tw_men_esi_oa_count = データ_tw_men_esi_oa.groupby(['ESI', 'bot', 'open_access_type']).count().reset_index()
データ_tw_men_esi_oa_count = データ_tw_men_esi_oa_count.pivot_table(values = 'DOI', index=['bot', 'ESI'], columns = 'open_access_type').reset_index()
データ_tw_men_esi_oa_count = データ_tw_men_esi_oa_count.merge(データ_tw_men_esi_count, how='inner', on=['bot', 'ESI'])

データ_tw_men_esi_oa_count
