==> 7_country.ipynb <==
# Bots in Science 🧪

## Libraries

import pandas as pd
from functions import bot_mentions

## 1. Correspondings' country

データ_cy = pd.read_csv('data/papers_country.tsv', sep='\t', encoding='UTF-8')
データ_cy.DOI = データ_cy.DOI.str.lower()
データ_cy = データ_cy.drop_duplicates()
データ_cy.shape

データ_tw_men = pd.read_csv('data/final_mentions_full_bots.tsv', sep='\t', encoding='UTF-8',
                         dtype={'Outlet or Author':str, 'External Mention ID':str})
データ_tw_men.DOI = データ_tw_men.DOI.str.lower()
データ_tw_men.shape

データ_tw_men[データ_tw_men.DOI.isin(データ_cy.DOI.tolist())].shape

データ_tw_men_cy = データ_tw_men.merge(データ_cy, how='inner', on='DOI')
データ_tw_men_cy.shape

データ_tw_men_cy = bot_mentions(データ_tw_men_cy, agg_by='country')
データ_tw_men_cy = データ_tw_men_cy.reset_index()
データ_tw_men_cy

データ_tw_men_cy.sort_values(by='mentions', ascending=False).head(20)

データ_tw_men_cy.to_csv('results/agg_bot_mentions_country.tsv', sep='\t', index=False)
